src/api/requirements.txt
alembic==1.2.1
apache-libcloud==3.0.0rc1
-e git+https://gitlab+deploy-token-12:1h2xVGJqWLnm1NnGGVVx@git.ascend.network/ascend/survey-js-parser.git@cd27c25f7692e6f2f0599dff7bf8d1f0035e7ce5#egg=ascend_survey_js_parser
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.8.0
blessed==1.17.0
blinker==1.4
certifi==2019.9.11
chardet==3.0.4
Click==7.0
coverage==5.0.3
factory-boy==2.12.0
Faker==4.0.3
Flask==1.1.1
Flask-Babel==1.0.0
Flask-Cloudy==1.1.0
Flask-Continuum==0.1.5
Flask-Cors==3.0.8
Flask-JWT-Extended==3.23.0
Flask-Mail==0.9.1
flask-marshmallow==0.12.0
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
GeoAlchemy2==0.7.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
lockfile==0.12.2
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-sqlalchemy==0.23.0
more-itertools==7.2.0
packaging==19.2
Pillow==7.1.2
pluggy==0.13.0
psycopg2-binary==2.8.3
py==1.8.0
-e git+https://gitlab+deploy-token-22:WLneXMF_zzEyefmPVcCp@git.ascend.network/ascend/py-metabase.git@10259fb59f468c9c2c392c96c00c4c478966ab50#egg=py_metabase
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.2.1
pytest-cov==2.8.1
pytest-mock==3.3.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
python-slugify==4.0.0
pytz==2019.3
PyYAML==3.13
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.8
SQLAlchemy-Continuum==1.3.9
SQLAlchemy-Utils==0.36.0
text-unidecode==1.3
urllib3==1.25.6
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
